Move flush to a plugin
Since usually we don't want the flush endpoint enabled, I think it
makes sense to move it out of the core into a built-in plugin for
two basic reasons:

* Simplify the application default config
* Uses the standard include interface
