+++ kinto/plugins/flush_endpoint.py
@@ -3,6 +3,7 @@ from pyramid.security import NO_PERMISSION_REQUIRED
 
 from kinto.events import ServerFlushed
 
+
 flush = Service(name='flush',
                 description='Clear database content',
                 path='/__flush__')
@@ -18,3 +19,13 @@ def flush_post(request):
 
     request.response.status = 202
     return {}
+
+
+def includeme(config):
+    config.add_api_capability(
+        "flush_endpoint",
+        description="The __flush__ endpoint can be used to remove all "
+                    "data from all backends.",
+        url="https://kinto.readthedocs.io/en/latest/configuration/"
+            "settings.html#activating-the-flush-endpoint")
+    config.add_cornice_service(flush)
