+++ tests/functional.ini
@@ -19,7 +19,6 @@ kinto.cache_url =
 # kinto.cache_max_size_bytes = 524288
 kinto.permission_backend = kinto.core.permission.memory
 kinto.permission_url =
-kinto.flush_endpoint_enabled = true
 
 #
 # Auth configuration.
@@ -32,7 +31,9 @@ multiauth.policies = basicauth
 #
 # Plugins
 #
-kinto.includes = kinto.plugins.default_bucket kinto.plugins.admin
+kinto.includes = kinto.plugins.default_bucket
+                 kinto.plugins.admin
+                 kinto.plugins.flush_endpoint
 
 #
 # Logging configuration
