+++ tests/test_views_hello.py
@@ -55,27 +55,6 @@ class HelloViewTest(BaseWebTest, unittest.TestCase):
         capabilities = resp.json['capabilities']
         self.assertNotIn('schema', capabilities)
 
-    def test_flush_capability_if_enabled(self):
-        settings = {'flush_endpoint_enabled': True}
-        app = self.make_app(settings=settings)
-        resp = app.get('/')
-        capabilities = resp.json['capabilities']
-        self.assertIn('flush_endpoint', capabilities)
-        expected = {
-            "description": "The __flush__ endpoint can be used to remove "
-                           "all data from all backends.",
-            "url": "https://kinto.readthedocs.io/en/latest/configuration/"
-                   "settings.html#activating-the-flush-endpoint"
-        }
-        self.assertEqual(expected, capabilities['flush_endpoint'])
-
-    def test_flush_capability_if_not_enabled(self):
-        settings = {'flush_endpoint_enabled': False}
-        app = self.make_app(settings=settings)
-        resp = app.get('/')
-        capabilities = resp.json['capabilities']
-        self.assertNotIn('flush_endpoint', capabilities)
-
     def test_permissions_capability_if_enabled(self):
         settings = {'experimental_permissions_endpoint': True}
         app = self.make_app(settings=settings)
